Added enemy Spawner and minor AI Director functionality
AI Director cycles through list of enemies and asks them to attack, however at the moment the cycle is started on player input for debugging purposes.

Will add a timer coroutine to give it some form of autonomy instead.

If no enemies are present in the list, the AIDirector simply does not do anything instead.

Every 1 second should be good for enemy attack management.