AI director has FINALLY been implemented. Although it is currently rather barebones and still full of bugs.
The few notable bugs are the fact that the enemies get removed from the list, during the iteration of the ForEach loop, which causes the ForEach loop to be calling a null reference.

A fix for this would be to implement a ListCleanup before the loop iteration, and instead of outright deleting the enemy prefab immediately when they "die". it is better to flag them as "IsDead" and use that boolean to denote them for corpse cleanup during every tick of the AiDirector() method.

//What should be done next few commits//
- Add an "IsDead" boolean to the EnemyAI script. This is so that the AIDirector can flag them for deletion later down the line by passing them into a list.
- Implement a "CorpseCleanup" function that allows the list to be modified AT THE START of the ForEach loop.
- Implement another Coroutine checker to allow some breathing room between each iteration of the loop/tick of the Ai Director. If the coroutine STILL has not been set to NULL after the end, that means AI Director tick() method should not be called just yet.