renamed some scripts ,also found a bug with the EnemyAI
The enemy when shifting into the HitStun state still stores a path (and still attempts to move on that path) during the hit stun animation.

This causes issues. Temporary fix would be to simply turn the navmeshAgent.speed to 0 for the duration of the hitstun state, and then go back to the default speed once it's done.