Updated SFX Manager to include 3 dictionaries. It's complicated so look at the description
Ignore all the sound files, they're just added in because I needed to do it anyways.

But the most important change is to the SFXManager and the GameEventSystem.

The SFXManager now has 3 dictionaries for player sounds, enemy sounds and environemnt sounds.
This dictionary interacts with the new GameEventSystem events by passing in a key to access the audioclip within the dictionary.

The dictionary uses the AudioClip's name as the key, and it uses a string key.

also there are now sounds for blocking attacks.